+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -97,8 +97,8 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
-- New :doc:`cppcoreguidelines-avoid-capture-this-with-capture-default
-  <clang-tidy/checks/cppcoreguidelines/avoid-capture-this-with-capture-default>` check.
+- New :doc:`cppcoreguidelines-avoid-capture-default-when-capturing-this
+  <clang-tidy/checks/cppcoreguidelines/avoid-capture-default-when-capturing-this>` check.
 
   Warns when lambda specify a capture default and capture ``this``. The check also
   offers fix-its.